// m-hub-frontend/src/app/enums/connection-type.enum.ts
/**
 * Enum for supported connection types / joining methods.
 */
export enum ConnectionType {
  conn_1 = 'Klebeverbindung',
  conn_2 = 'Schraubverbindung',
  conn_3 = 'Schweißverbindung',
  conn_4 = 'Nietverbindung',
  conn_5 = 'Nagelverbindung',
  conn_6 = 'Steckverbindung',
  conn_7 = 'Lötverbindung',
  conn_8 = 'Krimpverbindung',
  conn_9 = 'Pressverbindung',
  conn_10 = 'Keilverbindung'
}