// m-hub-frontend/src/app/enums/material-type.enum.ts
/**
 * Enum for supported materials.
 */

export enum MaterialType {
  mat_1 = 'Aluminium',
  mat_2 = 'Asphalt',
  mat_3 = 'Beton',
  mat_4 = 'Bitumen',
  mat_5 = 'Blähbeton',
  mat_6 = 'Blei',
  mat_7 = 'Diverse Kunststoffe',
  mat_8 = 'Eternit',
  mat_9 = 'Estrich',
  mat_10 = 'Fliesen',
  mat_11 = 'Fliesenkleber',
  mat_12 = 'Glas',
  mat_13 = 'Heraklith',
  mat_14 = 'Holz',
  mat_15 = 'Kautschuk',
  mat_16 = 'Keramik',
  mat_17 = 'Kupfer',
  mat_18 = 'Laminat',
  mat_19 = 'Linol',
  mat_20 = 'Messing',
  mat_21 = 'Mineralfaser',
  mat_22 = 'Mineralwolle',
  mat_23 = 'Mörtel',
  mat_24 = 'Naturstein',
  mat_25 = 'Papier',
  mat_26 = 'Putz',
  mat_27 = 'PVC',
  mat_28 = 'Rigips',
  mat_29 = 'Schlacke',
  mat_30 = 'Schüttung',
  mat_31 = 'Stahl',
  mat_32 = 'Steinzeug',
  mat_33 = 'Stroh',
  mat_34 = 'Styropor',
  mat_35 = 'Teppich',
  mat_36 = 'Terrazzo',
  mat_37 = 'Ytong',
  mat_38 = 'Ziegel'
}



// export enum MaterialType {
//   mat_1 = 'Ziegel',
//   mat_2 = 'Mörtel',
//   mat_3 = 'Beton',
//   mat_4 = 'Blähbeton',
//   mat_5 = 'Estrich',
//   mat_6 = 'Putz',
//   mat_7 = 'Stahl',
//   mat_8 = 'Heraklith',
//   mat_9 = 'Fliesen',
//   mat_10 = 'Fliesenkleber',
//   mat_11 = 'Rigips',
//   mat_12 = 'Holz',
//   mat_13 = 'Keramik',
//   mat_14 = 'Bitumen',
//   mat_15 = 'Glas',
//   mat_16 = 'Mineralwolle',
//   mat_17 = 'Styropor',
//   mat_18 = 'PVC',
//   mat_19 = 'Kupfer',
//   mat_20 = 'Aluminium',
//   mat_21 = 'Eternit',
//   mat_22 = 'Schlacke',
//   mat_23 = 'Blei',
//   mat_24 = 'Diverse Kunststoffe',
//   mat_25 = 'Stroh',
//   mat_26 = 'Naturstein',
//   mat_27 = 'Papier',
//   mat_28 = 'Messing',
//   mat_29 = 'Steinzeug',
//   mat_30 = 'Linoleum'
// } 